fs module implementation in full web is completed

+++ server/controller/technologies.js
@@ -1,3 +1,5 @@
+const fs = require("fs");
+const path = require("path");
 const connectDB = require("../database/connection");
 
 // get
@@ -82,20 +84,59 @@ const updateTechStack = async (req, res) => {
 const deleteTechStack = async (req, res) => {
   try {
     const { id } = req.params;
-    const Que = `DELETE FROM technologies WHERE id = ?`;
 
-    connectDB.query(Que, [id], (err) => {
+    // Retrieve the image filename from the database
+    const getImageQuery = `SELECT tech_photo FROM technologies WHERE id = ?`;
+    connectDB.query(getImageQuery, [id], (err, results) => {
       if (err) {
-        console.error(err.message);
-        return res
-          .status(500)
-          .json({ message: "error got from dleting tech stack" });
+        console.error(`Database error: ${err.message}`);
+        return res.status(500).json({ message: "Internal server error" });
       }
-      return res.sendStatus(200);
+
+      if (results.length > 0) {
+        const image = results[0].tech_photo;
+
+        // Delete the record from the database
+        const deleteQuery = `DELETE FROM technologies WHERE id = ?`;
+        connectDB.query(deleteQuery, [id], (err) => {
+          if (err) {
+            console.error(`Database error: ${err.message}`);
+            return res.status(500).json({ message: "Internal server error" });
+          }
+
+          // Construct the file path
+          const filePath = path.join(
+            __dirname,
+            "../../client/public/upload",
+            image
+          );
+          // Check if the file exists
+          fs.access(filePath, fs.constants.F_OK, (err) => {
+            if (err) {
+              console.error(`File does not exist: ${filePath}`);
+              return res.status(404).json({ message: "File not found" });
+            }
+
+            // Delete the image file from the local directory
+            fs.unlink(filePath, (err) => {
+              if (err) {
+                console.error(`Error deleting file: ${err.message}`);
+                return res
+                  .status(500)
+                  .json({ message: "Internal server error" });
+              }
+
+              return res.sendStatus(200);
+            });
+          });
+        });
+      } else {
+        return res.status(404).json({ message: "Record not found" });
+      }
     });
   } catch (error) {
-    console.error(error.message);
-    return res.status(500).json({ message: "internal server error" });
+    console.error(`Unexpected error: ${error.message}`);
+    return res.status(500).json({ message: "Internal server error" });
   }
 };
